Refactor BackendContext to handle nodes request/state + fix restart bug
